abstract/requirements.txt
google-cloud-storage>=2.10.0
requests>=2.31.0
pandas>=1.5.0
google-cloud-pubsub>=2.15.0
